+++ subprojects/common/analysis/src/main/java/hu/bme/mit/theta/analysis/algorithm/symbolic/fixedpoint/GeneralizedSaturationProvider.java
@@ -13,13 +13,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package hu.bme.mit.theta.analysis.algorithm.symbolic.fixpoint;
+package hu.bme.mit.theta.analysis.algorithm.symbolic.fixedpoint;
 
-import com.google.common.base.Preconditions;
 import com.koloboke.collect.set.hash.HashObjSets;
 import hu.bme.mit.delta.collections.IntObjCursor;
 import hu.bme.mit.delta.collections.IntObjMapView;
-import hu.bme.mit.delta.collections.impl.IntObjMapViews;
 import hu.bme.mit.delta.java.mdd.*;
 import hu.bme.mit.delta.java.mdd.impl.MddStructuralTemplate;
 import hu.bme.mit.theta.analysis.algorithm.symbolic.model.AbstractNextStateDescriptor;
@@ -29,7 +27,7 @@ import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.ToLongFunction;
 
-public final class CursorGeneralizedSaturationProvider implements MddTransformationProvider<AbstractNextStateDescriptor> {
+public final class GeneralizedSaturationProvider implements MddTransformationProvider<AbstractNextStateDescriptor> {
     public static boolean verbose = false;
 
     private MddVariableOrder variableOrder;
@@ -37,11 +35,11 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
     private final CacheManager<SaturationCache> cacheManager = new CacheManager<>(v -> new SaturationCache());
     private MddNode terminalZeroNode;
 
-    public CursorGeneralizedSaturationProvider(final MddVariableOrder variableOrder) {
+    public GeneralizedSaturationProvider(final MddVariableOrder variableOrder) {
         this(variableOrder, new LegacyRelationalProductProvider(variableOrder));
     }
 
-    public CursorGeneralizedSaturationProvider(
+    public GeneralizedSaturationProvider(
             final MddVariableOrder variableOrder, final RelationalProductProvider relProdProvider
     ) {
         this.variableOrder = variableOrder;
@@ -55,9 +53,7 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
             AbstractNextStateDescriptor nextStateRelation,
             MddVariableHandle highestAffectedVariable
     ) {
-        final MddHandle initialStates = relProdProvider.compute(((MddGraph<Boolean>) variableOrder.getMddGraph()).getHandleFor(true,
-                highestAffectedVariable
-        ), initializer, highestAffectedVariable);
+        final MddHandle initialStates = relProdProvider.compute(variableOrder.getMddGraph().getHandleForTop(), initializer, highestAffectedVariable);
 
         MddNode result;
 
@@ -101,16 +97,12 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
     public MddNode compute(
             final MddNode mddNode, final AbstractNextStateDescriptor nextState, final MddVariable mddVariable
     ) {
-        try (var nextStateCursor = nextState.rootCursor()) {
-            Preconditions.checkState(nextStateCursor.moveNext());
-            return saturate(mddNode, nextState, nextStateCursor, mddVariable, cacheManager.getCacheFor(mddVariable));
-        }
+        return saturate(mddNode, nextState, mddVariable, cacheManager.getCacheFor(mddVariable));
     }
 
     private MddNode saturate(
             final MddNode n,
             AbstractNextStateDescriptor d,
-            AbstractNextStateDescriptor.Cursor dCursor,
             MddVariable variable,
             CacheManager<SaturationCache>.CacheHolder cache
     ) {
@@ -135,32 +127,32 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
 
         final MddStateSpaceInfo stateSpaceInfo = new MddStateSpaceInfo(variable, n);
 
+//
 //		IntObjMapView<MddNode> satTemplate = new IntObjMapViews.Transforming<MddNode, MddNode>(n,
-//				(node, key) -> node == null ? null : terminalZeroToNull(saturate(node,
-//						d.getDiagonal(stateSpaceInfo).get(key),
-//						variable.getLower().orElse(null),
-//						cache.getLower()
-//				))
+//			(node, key) -> node == null ? null : terminalZeroToNull(saturate(node,
+//				d.getDiagonal(stateSpaceInfo).get(key),
+//				variable.getLower().orElse(null),
+//				cache.getLower()
+//			))
 //		);
 //
 //		MddNode nsat = variable.checkInNode(MddStructuralTemplate.of(satTemplate));
 
+
         MddUnsafeTemplateBuilder templateBuilder = JavaMddFactory.getDefault().createUnsafeTemplateBuilder();
 
         for (IntObjCursor<? extends MddNode> cFrom = n.cursor(); cFrom.moveNext(); ) {
-            try (var cTo = dCursor.valueCursor(cFrom.key())) {
-                MddNode s = saturate(cFrom.value(),
-                        cTo.moveTo(cFrom.key()) ? cTo.value() : AbstractNextStateDescriptor.terminalEmpty(),
-                        cTo,
-                        variable.getLower().orElse(null),
-                        cache.getLower()
-                );
 
-                templateBuilder.set(cFrom.key(),
-                        terminalZeroToNull(unionChildren(templateBuilder.get(cFrom.key()), s, variable))
-                );
+            MddNode s = saturate(cFrom.value(),
+                    d.getDiagonal(stateSpaceInfo).get(cFrom.key()),
+                    variable.getLower().orElse(null),
+                    cache.getLower()
+            );
 
-            }
+            templateBuilder.set(cFrom.key(),
+                    terminalZeroToNull(unionChildren(templateBuilder.get(cFrom.key()), s, variable))
+            );
+
         }
 
         MddNode nsat = variable.checkInNode(MddStructuralTemplate.of(templateBuilder.buildAndReset()));
@@ -170,38 +162,30 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
         do {
             changed = false;
 
-//			final Optional<Iterable<AbstractNextStateDescriptor>> splitNS = d.split();
-            final Optional<Iterable<AbstractNextStateDescriptor.Cursor>> splitNSCursor = dCursor.split();
-            if (splitNSCursor.isPresent()) {
-                for (AbstractNextStateDescriptor.Cursor dfireCursor : splitNSCursor.get()) {
+            final Optional<Iterable<AbstractNextStateDescriptor>> splitNS = d.split();
+            if (splitNS.isPresent()) {
+                for (AbstractNextStateDescriptor dfire : splitNS.get()) {
                     //System.out.println("Applying transition: " + dfire);
-                    if (dfireCursor.value().isLocallyIdentity(stateSpaceInfo)) {
+                    if (dfire.isLocallyIdentity(stateSpaceInfo)) {
                         continue;
                     }
-
-                    MddNode nfire = satFire(nsat, d, dfireCursor.value(), dfireCursor, variable, cache, stateSpaceInfo);
+                    MddNode nfire = satFire(nsat, d, dfire, variable, cache, stateSpaceInfo);
                     nfire = variable.union(nsat, nfire);
 
                     if (nfire != nsat) {
                         nsat = nfire;
                         changed = true;
                     }
-
-
                 }
             } else if (!d.isLocallyIdentity(stateSpaceInfo)) {
                 //System.out.println("Applying transition: " + d);
-//				try(var dCursor = d.rootCursor()){
-//					Preconditions.checkState(dCursor.moveNext());
-                MddNode nfire = satFire(nsat, d, d, dCursor, variable, cache, stateSpaceInfo);
+                MddNode nfire = satFire(nsat, d, d, variable, cache, stateSpaceInfo);
                 nfire = variable.union(nsat, nfire);
 
                 if (nfire != nsat) {
                     nsat = nfire;
                     changed = true;
                 }
-//				}
-
             }
         } while (changed);
 
@@ -224,7 +208,6 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
             MddNode n,
             AbstractNextStateDescriptor dsat,
             AbstractNextStateDescriptor dfire,
-            AbstractNextStateDescriptor.Cursor dfireCursor,
             MddVariable variable,
             CacheManager<SaturationCache>.CacheHolder cache,
             final MddStateSpaceInfo stateSpaceInfo
@@ -250,38 +233,42 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
 
         MddUnsafeTemplateBuilder templateBuilder = JavaMddFactory.getDefault().createUnsafeTemplateBuilder();
 
+//		final IntObjMapView<AbstractNextStateDescriptor> diagonal = dfire.getDiagonal(
+//				stateSpaceInfo);
+//		var c = diagonal.cursor();
+
+        var stateSpaceInfo2 = new MddStateSpaceInfo(variable, n);
+
         final IntObjMapView<IntObjMapView<AbstractNextStateDescriptor>> offDiagonal = dfire.getOffDiagonal(
-                stateSpaceInfo);
+                stateSpaceInfo2);
 
         for (IntObjCursor<? extends MddNode> cFrom = n.cursor(); cFrom.moveNext(); ) {
-            try (var cTo = dfireCursor.valueCursor(cFrom.key())) {
-                while (cTo.moveNext()) {
-                    if (cFrom.key() == cTo.key()) {
-                        continue;
-                    }
+            for (IntObjCursor<? extends AbstractNextStateDescriptor> cTo = offDiagonal.get(
+                    cFrom.key()).cursor(); cTo.moveNext(); ) {
+                if (cFrom.key() == cTo.key()) {
+                    continue;
+                }
 
-                    if (verbose) {
-                        System.out.println("Potential step: " + cFrom.key() + "->" + cTo.key());
-                    }
+                if (verbose) {
+                    System.out.println("Potential step: " + cFrom.key() + "->" + cTo.key());
+                }
 
-                    assert cFrom.value() != terminalZeroNode;
-                    assert cTo.value() != AbstractNextStateDescriptor.terminalEmpty();
+                assert cFrom.value() != terminalZeroNode;
+                assert cTo.value() != AbstractNextStateDescriptor.terminalEmpty();
 
-                    MddNode s = relProd(cFrom.value(),
-                            dsat.getDiagonal(stateSpaceInfo).get(cTo.key()),
-                            cTo.value(),
-                            cTo,
-                            variable.getLower().orElse(null),
-                            cache.getLower()
+                MddNode s = relProd(cFrom.value(),
+                        dsat.getDiagonal(stateSpaceInfo2).get(cTo.key()),
+                        cTo.value(),
+                        variable.getLower().orElse(null),
+                        cache.getLower()
+                );
+
+                if (s != terminalZeroNode) {
+                    confirm(variable, cTo.key());
+
+                    templateBuilder.set(cTo.key(),
+                            terminalZeroToNull(unionChildren(templateBuilder.get(cTo.key()), s, variable))
                     );
-
-                    if (s != terminalZeroNode) {
-                        confirm(variable, cTo.key());
-
-                        templateBuilder.set(cTo.key(),
-                                terminalZeroToNull(unionChildren(templateBuilder.get(cTo.key()), s, variable))
-                        );
-                    }
                 }
             }
         }
@@ -301,7 +288,6 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
             MddNode n,
             AbstractNextStateDescriptor dsat,
             AbstractNextStateDescriptor dfire,
-            AbstractNextStateDescriptor.Cursor dfireCursor,
             MddVariable variable,
             CacheManager<SaturationCache>.CacheHolder cache
     ) {
@@ -346,92 +332,62 @@ public final class CursorGeneralizedSaturationProvider implements MddTransformat
 
         for (IntObjCursor<? extends MddNode> cFrom = n.cursor(); cFrom.moveNext(); ) {
             // Identity step
-//			final AbstractNextStateDescriptor diagonalContinuation = diagonal.get(cFrom.key());
-//			if (!AbstractNextStateDescriptor.isNullOrEmpty(diagonalContinuation)) {
-//
-//				if (verbose) {
-//					System.out.println("Potential step: " + cFrom.key() + "->" + cFrom.key());
-//				}
-//
-//				MddNode s = relProd(cFrom.value(),
-//					dsat.getDiagonal(stateSpaceInfo).get(cFrom.key()),
-//					diagonalContinuation,
-//					variable.getLower().orElse(null),
-//					cache.getLower()
-//				);
-//
-//				if (s != terminalZeroNode) {
-//					// confirm(variable, cFrom.key());
-//
-//					templateBuilder.set(cFrom.key(),
-//						terminalZeroToNull(unionChildren(templateBuilder.get(cFrom.key()), s, variable))
-//					);
-//				}
-//			}
-            try (var cTo = dfireCursor.valueCursor(cFrom.key())) {
-                if (cTo.moveTo(cFrom.key())) {
+            final AbstractNextStateDescriptor diagonalContinuation = diagonal.get(cFrom.key());
+            if (!AbstractNextStateDescriptor.isNullOrEmpty(diagonalContinuation)) {
 
-                    if (verbose) {
-                        System.out.println("Potential step: " + cFrom.key() + "->" + cFrom.key());
-                    }
+                if (verbose) {
+                    System.out.println("Potential step: " + cFrom.key() + "->" + cFrom.key());
+                }
 
-                    MddNode s = relProd(cFrom.value(),
-                            dsat.getDiagonal(stateSpaceInfo).get(cFrom.key()),
-                            cTo.value(),
-                            cTo,
-                            variable.getLower().orElse(null),
-                            cache.getLower()
+                MddNode s = relProd(cFrom.value(),
+                        dsat.getDiagonal(stateSpaceInfo).get(cFrom.key()),
+                        diagonalContinuation,
+                        variable.getLower().orElse(null),
+                        cache.getLower()
+                );
+
+                if (s != terminalZeroNode) {
+                    // confirm(variable, cFrom.key());
+
+                    templateBuilder.set(cFrom.key(),
+                            terminalZeroToNull(unionChildren(templateBuilder.get(cFrom.key()), s, variable))
                     );
-
-                    if (s != terminalZeroNode) {
-                        // confirm(variable, cFrom.key());
-
-                        templateBuilder.set(cFrom.key(),
-                                terminalZeroToNull(unionChildren(templateBuilder.get(cFrom.key()), s, variable))
-                        );
-                    }
                 }
             }
 
-            try (var cTo = dfireCursor.valueCursor(cFrom.key())) {
-                while (cTo.moveNext()) {
-                    if (cFrom.key() == cTo.key()) {
-                        continue;
-                    }
+            for (IntObjCursor<? extends AbstractNextStateDescriptor> cTo = offDiagonal.get(cFrom.key()).cursor();
+                 cTo.moveNext(); ) {
+                if (cFrom.key() == cTo.key()) {
+                    continue;
+                }
 
-                    if (verbose) {
-                        System.out.println("Potential step: " + cFrom.key() + "->" + cTo.key());
-                    }
+                if (verbose) {
+                    System.out.println("Potential step: " + cFrom.key() + "->" + cTo.key());
+                }
 
-                    assert cFrom.value() != terminalZeroNode;
-                    assert cTo.value() != AbstractNextStateDescriptor.terminalEmpty();
+                assert cFrom.value() != terminalZeroNode;
+                assert cTo.value() != AbstractNextStateDescriptor.terminalEmpty();
 
-                    MddNode s = relProd(cFrom.value(),
-                            dsat.getDiagonal(stateSpaceInfo).get(cTo.key()),
-                            cTo.value(),
-                            cTo,
-                            variable.getLower().orElse(null),
-                            cache.getLower()
+                MddNode s = relProd(cFrom.value(),
+                        dsat.getDiagonal(stateSpaceInfo).get(cTo.key()),
+                        cTo.value(),
+                        variable.getLower().orElse(null),
+                        cache.getLower()
+                );
+
+                if (s != terminalZeroNode) {
+                    confirm(variable, cTo.key());
+
+                    templateBuilder.set(cTo.key(),
+                            terminalZeroToNull(unionChildren(templateBuilder.get(cTo.key()), s, variable))
                     );
-
-                    if (s != terminalZeroNode) {
-                        confirm(variable, cTo.key());
-
-                        templateBuilder.set(cTo.key(),
-                                terminalZeroToNull(unionChildren(templateBuilder.get(cTo.key()), s, variable))
-                        );
-                    }
                 }
             }
-
         }
 
         ret = variable.checkInNode(MddStructuralTemplate.of(templateBuilder.buildAndReset()));
 
-        try (var dsatCursor = dsat.rootCursor()) {
-            Preconditions.checkState(dsatCursor.moveNext());
-            ret = saturate(ret, dsat, dsatCursor, variable, cache);
-        }
+        ret = saturate(ret, dsat, variable, cache);
 
         cache.getCache().getRelProdCache().addToCache(n, dsat, dfire, ret);
